requirements.txt
discord.py==1.3.1
PyMySQL
pony
